Refactored code structure and add minor fixes
